Student score overwrite
