+++ common/lib/xmodule/xmodule/capa_module.py
@@ -331,10 +331,13 @@ class CapaDescriptor(CapaFields, RawDescriptor):
     new_lcp = module_attr('new_lcp')
     publish_grade = module_attr('publish_grade')
     rescore_problem = module_attr('rescore_problem')
+    rescore = module_attr('rescore')
     reset_problem = module_attr('reset_problem')
     save_problem = module_attr('save_problem')
+    set_score = module_attr('set_score')
     set_state_from_lcp = module_attr('set_state_from_lcp')
     should_show_submit_button = module_attr('should_show_submit_button')
     should_show_reset_button = module_attr('should_show_reset_button')
     should_show_save_button = module_attr('should_show_save_button')
     update_score = module_attr('update_score')
+    has_submitted_answer = module_attr('has_submitted_answer')
