+++ common/lib/xmodule/xmodule/js/src/capa/display.js
@@ -218,8 +218,9 @@
 
             if (curScore === undefined || totalScore === undefined) {
                 progress = '';
-            } else if (attemptsUsed === 0 || totalScore === 0) {
+            } else if ((attemptsUsed === 0 || totalScore === 0) && curScore === 0) {
                 // Render 'x point(s) possible' if student has not yet attempted question
+                // But if staff has overridden score to a non-zero number, show it
                 if (graded === 'True' && totalScore !== 0) {
                     progressTemplate = ngettext(
                         // Translators: %(num_points)s is the number of points possible (examples: 1, 3, 10).;
