+++ lms/djangoapps/instructor_task/api.py
@@ -12,9 +12,19 @@ import hashlib
 from celery.states import READY_STATES
 
 from xmodule.modulestore.django import modulestore
-
+from bulk_email.models import CourseEmail
+from certificates.models import CertificateGenerationHistory
+from lms.djangoapps.instructor_task.api_helper import (
+    check_arguments_for_rescoring,
+    check_arguments_for_overriding,
+    check_entrance_exam_problems_for_rescoring,
+    encode_entrance_exam_and_student_input,
+    encode_problem_and_student_input,
+    submit_task
+)
 from lms.djangoapps.instructor_task.models import InstructorTask
 from lms.djangoapps.instructor_task.tasks import (
+    override_problem_score,
     rescore_problem,
     reset_problem_attempts,
     delete_problem_state,
@@ -114,6 +124,28 @@ def submit_rescore_problem_for_student(request, usage_key, student, only_if_high
     task_class = rescore_problem
     task_input, task_key = encode_problem_and_student_input(usage_key, student)
     task_input.update({'only_if_higher': only_if_higher})
+    return submit_task(request, task_type, task_class, usage_key.course_key, task_input, task_key)
+
+
+def submit_override_score(request, usage_key, student, score):
+    """
+    Request a problem score override as a background task. Only
+    applicable to individual users.
+
+    The problem score will be overridden for the specified student only.
+    Parameters are the `course_id`, the `problem_url`, the `student` as
+    a User object, and the score override desired.
+    The url must specify the location of the problem, using i4x-type notation.
+
+    ItemNotFoundException is raised if the problem doesn't exist, or AlreadyRunningError
+    if this task is already running for this student, or NotImplementedError if
+    the problem is not a ScorableXBlock.
+    """
+    check_arguments_for_overriding(usage_key, score)
+    task_type = override_problem_score.__name__
+    task_class = override_problem_score
+    task_input, task_key = encode_problem_and_student_input(usage_key, student)
+    task_input['score'] = score
     return submit_task(request, task_type, task_class, usage_key.course_key, task_input, task_key)
