+++ lms/static/js/staff_debug_actions.js
@@ -19,12 +19,22 @@ var StaffDebug = (function() {
         return uname;
     };
 
+    var getScore = function(locationName) {
+        var sanitizedLocationName = sanitizeString(locationName);
+        var score = $('#sd_fs_' + sanitizedLocationName).val();
+        if (score === '') {
+            score = $('#sd_fs_' + sanitizedLocationName).attr('placeholder');
+        }
+        return score;
+    };
+
     var doInstructorDashAction = function(action) {
         var pdata = {
             problem_to_reset: action.location,
             unique_student_identifier: getUser(action.locationName),
             delete_module: action.delete_module,
-            only_if_higher: action.only_if_higher
+            only_if_higher: action.only_if_higher,
+            score: action.score
         };
         $.ajax({
             type: 'POST',
@@ -105,6 +115,17 @@ var StaffDebug = (function() {
         });
     };
 
+    var overrideScore = function(locname, location) {
+        this.doInstructorDashAction({
+            locationName: locname,
+            location: location,
+            method: 'override_problem_score',
+            success_msg: gettext('Successfully overrode problem score for {user}'),
+            error_msg: gettext('Could not override problem score for {user}.'),
+            score: getScore(locname)
+        });
+    };
+
     getCurrentUrl = function() {
         return window.location.pathname;
     };
@@ -114,12 +135,14 @@ var StaffDebug = (function() {
         deleteStudentState: deleteStudentState,
         rescore: rescore,
         rescoreIfHigher: rescoreIfHigher,
+        overrideScore: overrideScore,
 
         // export for testing
         doInstructorDashAction: doInstructorDashAction,
         getCurrentUrl: getCurrentUrl,
         getURL: getURL,
         getUser: getUser,
+        getScore: getScore,
         sanitizeString: sanitizeString
     }; })();
 
@@ -140,6 +163,11 @@ $(document).ready(function() {
     });
     $courseContent.on('click', '.staff-debug-rescore-if-higher', function() {
         StaffDebug.rescoreIfHigher($(this).parent().data('location-name'), $(this).parent().data('location'));
+        return false;
+    });
+
+    $courseContent.on('click', '.staff-debug-override-score', function() {
+        StaffDebug.overrideScore($(this).parent().data('location-name'), $(this).parent().data('location'));
         return false;
     });
 });
