+++ lms/static/sass/course/courseware/_courseware.scss
@@ -492,7 +492,7 @@ html.video-fullscreen {
         }
       }
 
-      .vert > .xblock-student_view {
+      .vert {
         @extend .clearfix;
         border-bottom: 1px solid #ddd;
         margin-bottom: ($baseline*0.75);
